+++ CreditManagementSystemFX/.idea/dataSources/9b052557-ab89-4029-a12e-537debc4eee9/storage_v2/_src_/database/CreditManagementSystem.i-HWkQ.meta
@@ -1,0 +1,1 @@
+#n:CreditManagementSystem
