+++ CreditManagementSystemFX/.idea/dataSources/9b052557-ab89-4029-a12e-537debc4eee9/storage_v2/_src_/database/CreditManagementSystem.i-HWkQ/schema/pg_catalog.0S1ZNQ.meta
@@ -1,0 +1,2 @@
+#n:pg_catalog
+!<md> [null, 0, null, null, -2147483648, -2147483648]
