+++ CreditManagementSystemFX/.idea/dataSources/9b052557-ab89-4029-a12e-537debc4eee9/storage_v2/_src_/database/CreditManagementSystem.i-HWkQ/schema/public.abK9xQ.meta
@@ -1,0 +1,2 @@
+#n:public
+!<md> [884, 0, null, null, -2147483648, -2147483648]
