+++ CreditManagementSystemFX/.idea/libraries/Maven__org_postgresql_postgresql_42_2_10_jre7.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.postgresql:postgresql:42.2.10.jre7">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/postgresql/postgresql/42.2.10.jre7/postgresql-42.2.10.jre7.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/postgresql/postgresql/42.2.10.jre7/postgresql-42.2.10.jre7-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/postgresql/postgresql/42.2.10.jre7/postgresql-42.2.10.jre7-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
