+++ CreditManagementSystemFX/pom.xml
@@ -20,6 +20,11 @@
             <artifactId>javafx-fxml</artifactId>
             <version>15.0.1</version>
         </dependency>
+        <dependency>
+            <groupId>org.postgresql</groupId>
+            <artifactId>postgresql</artifactId>
+            <version>42.2.10.jre7</version>
+        </dependency>
     </dependencies>
     <build>
         <plugins>
