+++ CreditManagementSystemFX/src/main/java/module-info.java
@@ -1,6 +1,8 @@
 module org.openjfx {
     requires javafx.controls;
     requires javafx.fxml;
+    requires java.sql;
+    requires postgresql;
 
     opens presentation to javafx.fxml;
     exports presentation;
