+++ CreditManagementSystemFX/src/main/java/persistence/PersistenceHandler.java
@@ -1,150 +1,38 @@
 package persistence;
 
-import java.io.*;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Scanner;
-import java.util.ArrayList;
-
-
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
 
 public class PersistenceHandler {
-    //Files representing a temporary DB
+    private static PersistenceHandler instance;
+    private Connection connection = null;
 
-    private File credit = new File("src/main/resources/credits.txt");
-    private File episode = new File("src/main/resources/episode.txt");
-    private File tvSeries = new File("src/main/resources/TvSeries.txt");
-    private File person = new File("src/main/resources/person.txt");
-    private File transmission = new File("src/main/resources/transmission.txt");
-    private File systemAdmin = new File("src/main/resources/systemadmin.txt");
-    private File producer = new File("src/main/resources/producer.txt");
-
-    public FileWriter fileWriter;
-
-    public PersistenceHandler() {
+    private PersistenceHandler(){
+        initializePostgresqlDatabase();
     }
 
-    //Contains the general write functionality that is used to write the files
-    private void write(File file, String s) {
-        if (!file.exists()) {
-            try {
-                file.createNewFile();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
+    public static PersistenceHandler getInstance(){
+        if (instance == null) {
+            instance = new PersistenceHandler();
         }
+        return instance;
+    }
 
+    private void initializePostgresqlDatabase() {
         try {
-            fileWriter = new FileWriter(file, true);
-            PrintWriter printWriter = new PrintWriter(fileWriter);
-            printWriter.println(s);
-            printWriter.close();
-
-        } catch (IOException e) {
-            e.printStackTrace();
+            DriverManager.registerDriver(new org.postgresql.Driver());
+            String url = "localhost";
+            int port = 5432;
+            connection = DriverManager.getConnection("jdbc:postgresql://" + url + ":" + port + "/" +
+                    DatabaseConfig.getDatabaseName(),
+                    DatabaseConfig.getUsername(),
+                    DatabaseConfig.getPassword());
+        } catch (SQLException | IllegalArgumentException ex) {
+            ex.printStackTrace(System.err);
+        } finally {
+            if (connection == null) System.exit(-1);
         }
-    }
-
-    public void writeCredit(String s) {
-        write(credit, s);
-    }
-
-    public void writePerson(String s) {
-        write(person, s);
-    }
-
-    public void writeTvSeries(String s) {
-        write(tvSeries, s);
-    }
-
-    public void writeTransmission(String s) {
-        write(transmission, s);
-    }
-
-    public void writeEpisode(String s) {
-        write(episode, s);
-    }
-
-    public void writeSystemAdmin(String s) {
-        write(systemAdmin, s);
-    }
-
-    public void writeProducer(String s) {
-        write(producer, s);
-    }
-
-    public ArrayList<String[]> readCredit() {
-        return read(credit);
-    }
-
-    public ArrayList<String[]> readEpisode() {
-        return read(episode);
-    }
-
-    public ArrayList<String[]> readPerson() {
-        return read(person);
-    }
-
-    public ArrayList<String[]> readTvSeries() {
-        return read(tvSeries);
-    }
-
-    public ArrayList<String[]> readTransmission() {
-        return read(transmission);
-    }
-
-    public ArrayList<String[]> readSystemAdmin() {
-        return read(systemAdmin);
-    }
-
-    public ArrayList<String[]> readProducer() {
-        return read(producer);
-    }
-
-    //Contains the general read functionality that is used to read the files
-    public ArrayList<String[]> read(File file) {
-        //
-        ArrayList<String[]> list = new ArrayList<>();
-
-        try {
-            Scanner scanner = new Scanner(file);
-            while(scanner.hasNextLine()){
-                String string = scanner.nextLine();
-                String[] data = string.split(";");
-                list.add(data);
-            }
-        } catch (FileNotFoundException e) {
-//            e.printStackTrace();
-        }
-        return list;
-    }
-
-    public File getCreditFile() {
-        return credit;
-    }
-
-    public File getEpisode() {
-        return episode;
-    }
-
-    public File getTvSeries() {
-        return tvSeries;
-    }
-
-    public File getPerson() {
-        return person;
-    }
-
-    public File getTransmission() {
-        return transmission;
-    }
-
-    public File getSystemAdmin() {
-        return systemAdmin;
-    }
-
-    public File getProducerFile() {
-        return producer;
     }
 
 }
